feat : Card를 감싸는 일급 컬렉션 추가
